+++ Rakefile
@@ -1,10 +1,10 @@
-if ENV['RACK_ENV'] != 'production'
-  require 'rspec/core/rake_task'
-
-  RSpec::Core::RakeTask.new :spec
-
-  task default: [:spec]
-end
+# if ENV['RACK_ENV'] != 'production'
+#   require 'rspec/core/rake_task'
+#
+#   RSpec::Core::RakeTask.new :spec
+#
+#   task default: [:spec]
+# end
 
 require 'data_mapper'
 require './app/app.rb'
