+++ app/models/user.rb
@@ -11,7 +11,6 @@ class User
   property :name, String
   property :username, String, unique: true, required: true
   property :phone, String
-  property :password_digest, String
   property :email, String, format: :email_address, required: true, unique: true
   property :password_digest, Text
   # property :password_token, Text
